Make tests more robust.
